Some Unicode objects were still converted to strings - therefore also bumped the genpy version.
While I was at it, the Python version is also written to the file.
Later, gencache should add this check to the ones it already
makes for a valid generated file (different Python versions will
cause the generated file to be different.  However, as the default
location for the generated files is version specific, this isnt a priority.